// src/store/userStore/type.d.ts
export interface UserStore {
  userId: string;
  groomName: string;
  groomPhone: string;
  groomAccount: string;
  brideName: string;
  bridePhone: string;
  brideAccount: string;
  weddingDate: string;
  weddingAddress: string;
  weddingDetailAddress: string;
  groomFatherName: string;
  groomFatherPhone: string;
  groomFatherAccount: string;
  groomMotherName: string;
  groomMotherPhone: string;
  groomMotherAccount: string;
  brideFatherName: string;
  brideFatherPhone: string;
  brideFatherAccount: string;
  brideMotherName: string;
  brideMotherPhone: string;
  brideMotherAccount: string;
  thumbnails: string[];
  galleries: IGallery[];
  setUser: (user: Partial<UserStore>) => void;
  clearUser: () => void;
  errors: Partial<Record<string, string[]>>;
  setErrors: (errors: Partial<Record<string, string[]>>) => void;
  clearErrors: () => void;
  isUser: boolean;
}